+++ Sources/Observability/Client/ObservabilityClientFactory.swift
@@ -87,15 +87,23 @@ public struct ObservabilityClientFactory {
                 autoInstrumentation.append(NetworkInstrumentationManager(options: options, tracer: decorator, session: sessionManager))
             }
             tracer = decorator
+            if options.autoInstrumentation.contains(.launchTimes) {
+                options.launchMeter.subscribe { statistics in
+                    for element in statistics {
+                        let span = decorator.startSpan(
+                            name: "AppStart",
+                            attributes: ["start.type": .string(element.launchType.description)],
+                            startTime: element.startTime
+                        )
+                        span.end(time: element.endTime)
+                    }
+                }
+            }
         } else {
             tracer = NoOpTracer()
         }
         
         let userInteractionManager = UserInteractionManager(options: options) { interaction in
-            Task {
-                await eventQueue.send(interaction)
-            }
-            
             // TODO: move to LD buffering
             if let span = interaction.span() {
                 tracer.startSpan(name: span.name, attributes: span.attributes)
@@ -161,7 +169,8 @@ public struct ObservabilityClientFactory {
             options: options,
             appLifecycleManager: appLifecycleManager,
             sessionManager: sessionManager,
-            transportService: transportService
+            transportService: transportService,
+            userInteractionManager: userInteractionManager
         )
         
         transportService.start()