+++ Sources/SessionReplay/Exporter/TouchInteraction+Payload.swift
@@ -1,9 +1,13 @@
 import Foundation
 import Common
+import Observability
 
 extension TouchInteraction: EventQueueItemPayload {
+    public var exporterClass: AnyClass {
+        SessionReplayExporter.self
+    }
+    
     public func cost() -> Int {
         300
     }
 }
-